GEOMESA-109 Refactor IndexSchema
* Broke up the IndexSchema file into separate files.
* Changed the STII's use of an IndexSchema to a IndexEntryDecoder.
* Pushed everything down into the IndexQueryPlanner in preparation for work on it.
