+++ geomesa-core/src/main/scala/geomesa/core/index/IndexEntry.scala
@@ -1,0 +1,111 @@
+package geomesa.core.index
+
+import com.vividsolutions.jts.geom.Geometry
+import geomesa.core.data.SimpleFeatureEncoder
+import geomesa.core.iterators.AttributeAggregator
+import geomesa.utils.geohash.GeohashUtils
+import org.apache.accumulo.core.data.{Value, Key}
+import org.apache.hadoop.io.Text
+import org.geotools.feature.simple.SimpleFeatureBuilder
+import org.joda.time.DateTime
+import org.opengis.feature.simple.SimpleFeature
+import scala.collection.JavaConversions._
+
+object IndexEntry {
+
+  implicit class IndexEntrySFT(sf: SimpleFeature) {
+    lazy val userData = sf.getFeatureType.getUserData
+    lazy val dtgStartField = userData.getOrElse(SF_PROPERTY_START_TIME, SF_PROPERTY_START_TIME).asInstanceOf[String]
+    lazy val dtgEndField = userData.getOrElse(SF_PROPERTY_END_TIME, SF_PROPERTY_END_TIME).asInstanceOf[String]
+
+    lazy val sid = sf.getID
+    lazy val gh = GeohashUtils.reconstructGeohashFromGeometry(geometry)
+    def geometry = sf.getDefaultGeometry.asInstanceOf[Geometry]
+    def setGeometry(geometry: Geometry) {
+      sf.setDefaultGeometry(geometry)
+    }
+
+    private def getTime(attr: String) = sf.getAttribute(attr).asInstanceOf[java.util.Date]
+    def startTime = getTime(dtgStartField)
+    def endTime   = getTime(dtgEndField)
+    lazy val dt   = Option(startTime).map { d => new DateTime(d) }
+
+    private def setTime(attr: String, time: DateTime) =
+      sf.setAttribute(attr, if (time == null) null else time.toDate)
+
+    def setStartTime(time: DateTime) = setTime(dtgStartField, time)
+    def setEndTime(time: DateTime)   = setTime(dtgEndField, time)
+  }
+
+}
+
+case class IndexEncoder(rowf: TextFormatter[SimpleFeature],
+                        cff: TextFormatter[SimpleFeature],
+                        cqf: TextFormatter[SimpleFeature],
+                        featureEncoder: SimpleFeatureEncoder) {
+
+  import GeohashUtils._
+  import IndexEntry._
+
+  val formats = Array(rowf,cff,cqf)
+
+  // the resolutions are valid for decomposed objects are all 5-bit boundaries
+  // between 5-bits and 35-bits (inclusive)
+  lazy val decomposableResolutions: ResolutionRange = new ResolutionRange(0, 35, 5)
+
+  // the maximum number of sub-units into which a geometry may be decomposed
+  lazy val maximumDecompositions: Int = 5
+
+  def encode(entryToEncode: SimpleFeature): List[KeyValuePair] = {
+    // decompose non-point geometries into multiple index entries
+    // (a point will return a single GeoHash at the maximum allowable resolution)
+    val geohashes =
+      decomposeGeometry(entryToEncode.geometry, maximumDecompositions, decomposableResolutions)
+
+    val entries = geohashes.map { gh =>
+      val copy = SimpleFeatureBuilder.copy(entryToEncode)
+      val geom = getGeohashGeom(gh)
+      copy.setDefaultGeometry(geom)
+      copy
+    }
+
+    // remember the resulting index-entries
+    val keys = entries.map { entry =>
+      val Array(r, cf, cq) = formats.map { _.format(entry) }
+      new Key(r, cf, cq, entry.dt.map(_.getMillis).getOrElse(DateTime.now().getMillis))
+    }
+    val rowIDs = keys.map(_.getRow)
+    val id = new Text(entryToEncode.sid)
+
+    val indexValue = IndexSchema.encodeIndexValue(entryToEncode)
+    val iv = new Value(indexValue)
+    // the index entries are (key, FID) pairs
+    val indexEntries = keys.map { k => (k, iv) }
+
+    // the (single) data value is the encoded (serialized-to-string) SimpleFeature
+    val dataValue = featureEncoder.encode(entryToEncode)
+
+    // data entries are stored separately (and independently) from the index entries;
+    // each attribute gets its own data row (though currently, we use only one attribute
+    // that represents the entire, encoded feature)
+    val dataEntries = rowIDs.map { rowID =>
+      val key = new Key(rowID, id, AttributeAggregator.SIMPLE_FEATURE_ATTRIBUTE_NAME_TEXT)
+      (key, dataValue)
+    }
+
+    (indexEntries ++ dataEntries).toList
+  }
+
+}
+
+case class IndexEntryDecoder(ghDecoder: GeohashDecoder,
+                             dtDecoder: Option[DateDecoder]) {
+
+  import GeohashUtils._
+
+  def decode(key: Key): SimpleFeature = {
+    val gh = getGeohashGeom(ghDecoder.decode(key))
+    val dt = dtDecoder.map(_.decode(key))
+    SimpleFeatureBuilder.build(indexSFT, List(gh, dt), "")
+  }
+}
